+++ src/main/java/domain/board/Board.java
@@ -1,7 +1,6 @@
-package domain;
+package domain.board;
 
 import domain.piece.Piece;
-
 import java.util.Map;
 
 public class Board {
