+++ src/main/java/domain/board/Edge.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 public class Edge {
