+++ src/main/java/domain/piece/Cha.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Cha implements Piece {
 
     private final Team team;
