+++ src/main/java/domain/piece/Ma.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Ma implements Piece {
 
     private final Team team;
