+++ src/main/java/domain/piece/Po.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Po implements Piece {
 
     private final Team team;
