+++ src/main/java/domain/piece/Sa.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Sa implements Piece {
 
     private final Team team;
