+++ src/main/java/domain/piece/Sang.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Sang implements Piece {
 
     private final Team team;
