+++ src/main/java/domain/piece/Team.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.piece;
 
 public enum Team {
