+++ src/main/java/domain/piece/Wang.java
@@ -1,7 +1,5 @@
 package domain.piece;
 
-import domain.Team;
-
 public class Wang implements Piece {
 
     private final Team team;
