+++ src/test/java/domain/PointTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.board.Point;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
